requirements.txt
httpx==0.27.2
playwright==1.46.0
jinja2==3.1.4
pillow==10.4.0
imagehash==4.3.1
tldextract==5.1.2
idna==3.7
certifi==2024.8.30
